requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==3.0.0
Django==4.1.3
django-allauth==0.52.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
install==1.3.5
oauthlib==3.2.2
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
